chore(backend): configure running e2e tests properly
