+++ packages/backend/.env.test
@@ -1,0 +1,2 @@
+NODE_ENV=test
+DATABASE_URL=postgresql://postgres:password@localhost:5432/ddd-forum
