+++ packages/backend/package.json
@@ -16,7 +16,7 @@
     "lint": "eslint .",
     "test": "jest",
     "test:dev": "jest --watchAll",
-    "test:e2e": "jest -c jest.config.js --verbose tests/e2e/ --runInBand --detectOpenHandles"
+    "test:e2e": "jest -c jest.config.e2e.ts --detectOpenHandles"
   },
   "prisma": {
     "seed": "ts-node prisma/seed.ts"
