+++ packages/backend/prepareEnv.ts
@@ -11,42 +11,18 @@
  */
 
 import { execSync } from 'child_process';
-import dotenv from 'dotenv';
-import fs from 'fs';
-import { cleanEnv, str } from 'envalid';
 import * as path from 'path';
+import { configureEnv } from './scripts/configureEnv';
 
-const validateEnv = (envFile: string) => {
-  cleanEnv(
-    {
-      ...dotenv.parse(fs.readFileSync(envFile)),
-    },
-    {
-      DATABASE_URL: str(),
-      NODE_ENV: str({ choices: ['development', 'test', 'production', 'staging'] }),
-    }
-  );
-};
-
-export const prepareEnv = (): void => {
-  const env = process.env.NODE_ENV || 'development';
+const prepareEnv = (): void => {
   const packageRoot = path.resolve(__dirname);
   const execParams = {
     cwd: packageRoot,
     stdio: 'inherit',
   } as const;
   const script = process.argv.splice(2).join(' ');
-  const envFile = `.env.${env}`;
-  const envFileExists = fs.existsSync(envFile);
 
-  if (envFileExists) {
-    validateEnv(envFile);
-    console.log(`Preparing dev environment using ${envFile}`);
-    execSync(`dotenv -e ${envFile} -- ${script}`, execParams);
-    return;
-  }
-
-  console.log(`Running ${script} in ${process.env.NODE_ENV} mode without loading from env file.`);
+  configureEnv();
   execSync(`${script}`, execParams);
 };
