+++ packages/backend/tests/support/globalDevEnvTestSetup.ts
@@ -1,0 +1,1 @@
+export { configureEnv as default } from '../../scripts/configureEnv';
